# backend/config/openapi.yaml
openapi: 3.0.4
info:
  title: Jmich Airlines API
  description: API for search and purchase cheap flight tickets
  version: 1.9.3
servers:
  - url: http://localhost:8000
    description: Development server
...
paths:
...
  /booking/{id}:
    get:
      tags: [Booking]
      summary: Get booking by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
    delete:
      tags: [Booking]
      summary: Delete booking
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
  /booking/:
    post:
      tags: [Booking]
      summary: Create new booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [flight_id, seats]
              properties:
                flight_id:
                  type: integer
                seats:
                  type: integer
                  minimum: 1
                baggage_weight:
                  type: number
                  minimum: 0
                code:
                  type: string
      responses:
        "201":
          description: Booking created
    get:
      tags: [Booking]
      summary: Get user's bookings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
...
  /booking/all:
    get:
      tags: [Booking]
      summary: Get all bookings (paginated)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Success
...
  /flight/search:
    post:
      tags: [Flight]
      summary: Search flights
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                departure_airport_id:
                  type: integer
                arrival_airport_id:
                  type: integer
                departure_date:
                  type: string
                  format: date
                seats:
                  type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightSearchResponse"
  /flight/{id}:
    get:
      tags: [Flight]
      summary: Get flight by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightResponse"
    put:
      tags: [Flight]
      summary: Update flight
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [
                  departure_airport_id,
                  arrival_airport_id,
                  company_id,
                  departure_date,
                  arrival_date,
                  route_code,
                  price,
                  seats_available,
                ]
              properties:
                departure_airport_id:
                  type: integer
                arrival_airport_id:
                  type: integer
                company_id:
                  type: integer
                departure_date:
                  type: string
                  format: date-time
                arrival_date:
                  type: string
                  format: date-time
                route_code:
                  type: string
                  pattern: "^[A-Z][0-9]{3}$"
                price:
                  type: number
                  minimum: 1
                seats_available:
                  type: integer
                  minimum: 0
      responses:
        "200":
          description: Success
    delete:
      tags: [Flight]
      summary: Delete flight
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
  /flight/:
    post:
      tags: [Flight]
      summary: Add new flight
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [
                  departure_airport_id,
                  arrival_airport_id,
                  company_id,
                  departure_date,
                  arrival_date,
                  route_code,
                  price,
                  seats_available,
                ]
              properties:
                departure_airport_id:
                  type: integer
                arrival_airport_id:
                  type: integer
                company_id:
                  type: integer
                departure_date:
                  type: string
                  format: date-time
                arrival_date:
                  type: string
                  format: date-time
                route_code:
                  type: string
                  pattern: "^[A-Z][0-9]{3}$"
                price:
                  type: number
                  minimum: 1
                seats_available:
                  type: integer
                  minimum: 0
      responses:
        "201":
          description: Flight created
  /flight/{id}/status:
    put:
      tags: [Flight]
      summary: Update flight status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
  /flight/all/{status}:
    get:
      tags: [Flight]
      summary: Get all flights (paginated)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: status
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Success
  /flight/random:
    post:
      tags: [Flight]
      summary: Add random flights
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Random flights added
...
components:
  schemas:
    BookingResponse:
      type: object
      properties:
        status:
          type: string
        bookings:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              user_id:
                type: integer
              baggage_weight:
                type: number
              qr_code:
                type: string
              created:
                type: string
                format: date-time
              seats:
                type: integer
              cost:
                type: number
              status:
                type: string
              flight:
                $ref: "#/components/schemas/FlightDetails"
    FlightDetails:
      type: object
      properties:
        id:
          type: integer
        route_code:
          type: string
        departure_city:
          $ref: "#/components/schemas/CityWithAirport"
        arrival_city:
          $ref: "#/components/schemas/CityWithAirport"
        company:
          $ref: "#/components/schemas/CompanyWithRule"
        departure_date:
          type: string
          format: date-time
        arrival_date:
          type: string
          format: date-time
        price:
          type: number
        seats_available:
          type: integer
        status:
          type: string
    CityWithAirport:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
        airport:
          $ref: "#/components/schemas/AirportDetails"
    AirportDetails:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
    CompanyWithRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logo:
          type: string
        baggage_rule:
          $ref: "#/components/schemas/BaggageRule"
    BaggageRule:
      type: object
      properties:
        max_free_weight:
          type: number
        price_per_kg:
          type: number
    FlightSearchResponse:
      type: object
      properties:
        status:
          type: string
        flights:
          type: array
          items:
            $ref: "#/components/schemas/FlightDetails"
    FlightResponse:
      type: object
      properties:
        status:
          type: string
        flight:
          $ref: "#/components/schemas/FlightDetails"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT